clip-dinoiser/requirements.txt
cityscapesscripts==2.2.4
ftfy==6.3.1
hydra-core==1.3.2
matplotlib==3.9.2
mmengine==0.10.5
mmsegmentation==0.27.0
numpy==1.26.4
omegaconf==2.3.0
open_clip_torch==2.24.0
opencv_python==4.10.0.84
Pillow==10.4.0
regex==2024.9.11
scipy==1.13.1
termcolor==2.5.0
tqdm==4.65.2
